Enforce comparison
a non-empty array would equal true if the comparison is "==". This would mean that if the validation has errors, the validation would pass as success. A non-empty array would not equal TRUE if the comparison is "===". This would mean if the validation has errors, the validation should not pass as success.